+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 ### Fixed
 - Replaced `sketchy-null`
+- Fixed type for event in `onError` and `onSubmit`
 
 ## [0.2.0] - 2019-09-22
 ### Added